Fix compiler warnings: old-style-cast, shadow, and sign-conversion
- Fixed all 2 [-Wold-style-cast] warnings by replacing C-style casts with static_cast
- Fixed all 8 [-Wshadow] warnings by removing duplicate type aliases and renaming lambda parameters
- Fixed all 528 [-Wsign-conversion] warnings by adding explicit static_cast conversions

Changes:
- compressed_graph.hpp: Cast max() result to vertex_id_type
- dynamic_graph.hpp: Cast e.source_id when indexing vertices_ array
- Test files: Cast int loop variables to size_t for array indexing
- Test files: Cast std::ranges::distance results from long int to size_t

Total warnings reduced from 1,039 to 8 (99.2% reduction)

+++ include/graph/container/dynamic_graph.hpp
@@ -1366,7 +1366,7 @@ public: // Load operations
               throw std::runtime_error("source id exceeds the number of vertices in load_edges");
             if (static_cast<size_t>(e.target_id) >= vertices_.size())
               throw std::runtime_error("target id exceeds the number of vertices in load_edges");
-            auto& vertex_edges = vertices_[e.source_id].edges();
+            auto& vertex_edges = vertices_[static_cast<size_t>(e.source_id)].edges();
             if constexpr (Sourced) {
               if constexpr (is_void_v<EV>) {
                 emplace_edge(vertex_edges, e.target_id, edge_type(e.source_id, e.target_id));
@@ -1396,7 +1396,7 @@ public: // Load operations
           throw std::runtime_error("source id exceeds the number of vertices in load_edges");
         if (static_cast<size_t>(e.target_id) >= vertices_.size())
           throw std::runtime_error("target id exceeds the number of vertices in load_edges");
-        auto& vertex_edges = vertices_[e.source_id].edges();
+        auto& vertex_edges = vertices_[static_cast<size_t>(e.source_id)].edges();
         if constexpr (Sourced) {
           if constexpr (is_void_v<EV>) {
             emplace_edge(vertex_edges, e.target_id, edge_type(e.source_id, e.target_id));
